Move two query files
