+++ frontend/src/components/CourseModal/RatingsGraph.tsx
@@ -4,7 +4,7 @@ import styled from 'styled-components';
 import styles from './RatingsGraph.module.css';
 import { TextComponent } from '../StyledComponents';
 
-import type { graphLabels } from '../../queries/Constants';
+import type { graphLabels } from '../../utilities/constants';
 
 const StyledLabel = styled.p`
   font-size: 10px !important;
