+++ frontend/src/contexts/worksheetContext.tsx
@@ -11,9 +11,8 @@ import {
   useSessionStorageState,
 } from '../utilities/browserStorage';
 import { CUR_SEASON } from '../config';
-import { useFerry } from './ferryContext';
+import { useFerry, useWorksheetInfo } from './ferryContext';
 import { toSeasonString } from '../utilities/course';
-import { useWorksheetInfo } from '../queries/GetWorksheetListings';
 import { useUser, type Worksheet } from './userContext';
 import type { Season, Listing } from '../utilities/common';
 import type { OptType } from './searchContext';
