+++ frontend/src/utilities/constants.ts
@@ -1,6 +1,6 @@
 import { expectType, type TypeOf } from 'ts-expect';
 import chroma from 'chroma-js';
-import type { Listing } from '../utilities/common';
+import type { Listing } from './common';
 
 // Phrases for search speed [50 character limit]
 export const searchSpeed = {
